requirements.txt
pandas
langchain_openai
python-dotenv
langchain-pinecone
langchain-community
jq
langchain-google-genai
fastapi
aiohttp
uvicorn